+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,18 @@
 # ai
 
+## 5.0.115
+
+### Patch Changes
+
+- Updated dependencies [1bcc2d7]
+  - @ai-sdk/gateway@2.0.22
+
+## 5.0.114
+
+### Patch Changes
+
+- fc633bd: fix header loss when statusText is undefined in writeHead
+
 ## 5.0.113
 
 ### Patch Changes