+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/google-vertex
 
+## 3.0.94
+
+### Patch Changes
+
+- 3ea80e7: feat(provider/google): add enterpriseWebSearch tool
+- Updated dependencies [3ea80e7]
+  - @ai-sdk/google@2.0.49
+
+## 3.0.93
+
+### Patch Changes
+
+- 1bcc2d7: feat(google): gemini-3-flash
+- Updated dependencies [1bcc2d7]
+  - @ai-sdk/google@2.0.48
+
 ## 3.0.92
 
 ### Patch Changes