+++ packages/google/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/google
 
+## 2.0.49
+
+### Patch Changes
+
+- 3ea80e7: feat(provider/google): add enterpriseWebSearch tool
+
+## 2.0.48
+
+### Patch Changes
+
+- 1bcc2d7: feat(google): gemini-3-flash
+
 ## 2.0.47
 
 ### Patch Changes