+++ packages/react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/react
 
+## 2.0.117
+
+### Patch Changes
+
+- ai@5.0.115
+
+## 2.0.116
+
+### Patch Changes
+
+- Updated dependencies [fc633bd]
+  - ai@5.0.114
+
 ## 2.0.115
 
 ### Patch Changes