+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/svelte
 
+## 3.0.115
+
+### Patch Changes
+
+- ai@5.0.115
+
+## 3.0.114
+
+### Patch Changes
+
+- Updated dependencies [fc633bd]
+  - ai@5.0.114
+
 ## 3.0.113
 
 ### Patch Changes